Extract spec parsers into SpecParsing
Move the command-line spec parsers (secret, ulimit, label, driver option,
filter) and the scalar value parsers (signal, timestamp, format, inspect
type, memory size, duration, nano-CPUs) out of ArgumentValidation into a
dedicated SpecParsing translation unit. The multi-field parsers now share a
common SplitKeyValue helper. The Validate* wrappers remain in
ArgumentValidation and forward to the moved parsers.
