+++ src/windows/wslc/arguments/ArgumentValidation.h
@@ -16,6 +16,7 @@ Abstract:
 #include "Exceptions.h"
 #include "ContainerModel.h"
 #include "InspectModel.h"
+#include "SpecParsing.h"
 #include <string>
 #include <tuple>
 #include <vector>
@@ -25,10 +26,6 @@ Abstract:
 #include "Localization.h"
 
 using namespace wsl::windows::wslc::models;
-
-namespace wsl::windows::wslc::services {
-struct BuildSecret;
-}
 
 namespace wsl::windows::wslc::validation {
 
@@ -65,36 +62,21 @@ T GetIntegerFromString(
 }
 
 void ValidateWSLCSignalFromString(const std::vector<std::wstring>& values, const std::wstring& argName);
-WSLCSignal GetWSLCSignalFromString(const std::wstring& input, const std::wstring& argName = {});
 
 void ValidateMemorySize(const std::vector<std::wstring>& values, const std::wstring& argName);
-int64_t GetMemorySizeFromString(const std::wstring& input, const std::wstring& argName = {});
 
 void ValidateDuration(const std::vector<std::wstring>& values, const std::wstring& argName);
-int64_t GetDurationNanosFromString(const std::wstring& input, const std::wstring& argName = {});
 
 void ValidateTimestamp(const std::vector<std::wstring>& values, const std::wstring& argName);
-ULONGLONG GetTimestampFromString(const std::wstring& value, const std::wstring& argName = {});
 void ValidateNanoCpus(const std::vector<std::wstring>& values, const std::wstring& argName);
-int64_t GetNanoCpusFromString(const std::wstring& input, const std::wstring& argName = {});
 
 void ValidateUlimit(const std::vector<std::wstring>& values, const std::wstring& argName);
-std::tuple<std::string, int64_t, int64_t> ParseUlimit(const std::wstring& input, const std::wstring& argName = {});
 
 void ValidateFormatTypeFromString(const std::vector<std::wstring>& values, const std::wstring& argName);
-FormatType GetFormatTypeFromString(const std::wstring& input, const std::wstring& argName = {});
-
-InspectType GetInspectTypeFromString(const std::wstring& input, const std::wstring& argName);
 
 void ValidateGpus(const std::vector<std::wstring>& values, const std::wstring& argName);
 
-services::BuildSecret ParseSecretSpec(const std::wstring& spec);
-
 void ValidateVolumeMount(const std::vector<std::wstring>& values);
 void ValidateFilter(const std::vector<std::wstring>& values);
-
-std::pair<std::string, std::string> ParseLabel(const std::wstring& value);
-std::pair<std::string, std::string> ParseDriverOption(const std::wstring& value);
-std::pair<std::string, std::string> ParseFilter(const std::wstring& value);
 
 } // namespace wsl::windows::wslc::validation
